+++ src/InverseCdf.h
@@ -4,26 +4,26 @@
 
 class InverseCdf
 {
-	double cdf_values_[CDF_RES + 1];
+	float cdf_values_[CDF_RES + 1];
 
 public:
 
-	void set_neg_log(double start_value);
+	void set_neg_log(float start_value);
 
 	void assign_exponent();
 
-	void assign_exponent(double value);
+	void assign_exponent(float value);
 
-	unsigned short int choose(double Mean, int tn, double timesteps_per_day);
+	unsigned short int choose(float Mean, int tn, float timesteps_per_day);
 
 	// Getter
-	double* get_values()
+	float* get_values()
 	{
 		return cdf_values_;
 	}
 
 	// Overloading [] operator to access elements in array style 
-	double& operator[](int i)
+	float& operator[](int i)
 	{
 		return cdf_values_[i];
 	}
